backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.14.2
shapely==2.0.2

# Cache
redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7

# HTTP
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.1.4
numpy==1.26.3

# ML
xgboost==2.0.3
onnxruntime==1.16.3
shap==0.44.1
scikit-learn==1.4.0

# Orchestration - lightweight scheduler
apscheduler==3.10.4

# Storage
minio==7.2.3

# Integrations
twilio==8.12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
